Add global error page and update dynamic route params
