+++ src/app/dashboard/collectives/[collectiveId]/settings/page.tsx
@@ -5,9 +5,9 @@ import EditCollectiveSettingsForm from "./EditCollectiveSettingsForm"; // Client
 export default async function CollectiveSettingsPage({
   params,
 }: {
-  params: { collectiveId: string };
+  params: Promise<{ collectiveId: string }>;
 }) {
-  const { collectiveId } = params;
+  const { collectiveId } = await params;
   const supabase = await createServerSupabaseClient();
 
   const {
